계층형 댓글 구현하기

+++ src/main/java/com/peeerr/climbing/service/PostService.java
@@ -1,15 +1,17 @@
 package com.peeerr.climbing.service;
 
-import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.exception.constant.ErrorMessage;
 import com.peeerr.climbing.domain.category.Category;
 import com.peeerr.climbing.domain.category.CategoryRepository;
 import com.peeerr.climbing.domain.post.Post;
 import com.peeerr.climbing.domain.post.PostRepository;
+import com.peeerr.climbing.domain.user.Member;
 import com.peeerr.climbing.dto.post.request.PostCreateRequest;
 import com.peeerr.climbing.dto.post.request.PostEditRequest;
 import com.peeerr.climbing.dto.post.response.PostResponse;
+import com.peeerr.climbing.dto.post.response.PostWithCommentsResponse;
+import com.peeerr.climbing.exception.constant.ErrorMessage;
 import com.peeerr.climbing.exception.ex.EntityNotFoundException;
+import com.peeerr.climbing.exception.ex.UnauthorizedAccessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -45,6 +47,14 @@ public class PostService {
         return postResponse;
     }
 
+    @Transactional(readOnly = true)
+    public PostWithCommentsResponse getPostWithComments(Long postId) {
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new EntityNotFoundException(ErrorMessage.POST_NOT_FOUND));
+
+        return PostWithCommentsResponse.from(post);
+    }
+
     @Transactional
     public PostResponse addPost(PostCreateRequest postCreateRequest, Member member) {
         Post post = Post.builder()
@@ -60,9 +70,13 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponse editPost(Long postId, PostEditRequest postEditRequest) {
+    public PostResponse editPost(Long postId, PostEditRequest postEditRequest, Long loginId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorMessage.POST_NOT_FOUND));
+
+        if (!post.getMember().getId().equals(loginId)) {
+            throw new UnauthorizedAccessException(ErrorMessage.NO_ACCESS_PERMISSION);
+        }
 
         post.changeTitle(postEditRequest.getTitle());
         post.changeContent(postEditRequest.getContent());
@@ -72,9 +86,13 @@ public class PostService {
     }
 
     @Transactional
-    public void removePost(Long postId) {
+    public void removePost(Long postId, Long loginId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorMessage.POST_NOT_FOUND));
+
+        if (!post.getMember().getId().equals(loginId)) {
+            throw new UnauthorizedAccessException(ErrorMessage.NO_ACCESS_PERMISSION);
+        }
 
         postRepository.delete(post);
     }
